requirements.txt
python-telegram-bot==20.7
watchdog==3.0.0
asyncio